Bug Fix - 2 successful

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -1,8 +1,10 @@
 package ImageHoster.controller;
 
+import ImageHoster.model.Comment;
 import ImageHoster.model.Image;
 import ImageHoster.model.Tag;
 import ImageHoster.model.User;
+import ImageHoster.service.CommentService;
 import ImageHoster.service.ImageService;
 import ImageHoster.service.TagService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -26,6 +28,9 @@ public class ImageController {
 
     @Autowired
     private TagService tagService;
+
+    @Autowired
+    private CommentService commentService;
 
     //This method displays all the images in the user home page after successful login
     @RequestMapping("images")
@@ -51,8 +56,12 @@ public class ImageController {
     public String showImage(@PathVariable("imageId") Integer imageId, @PathVariable("title") String title, Model model) {
         //Image image = imageService.getImageByTitle(title);
         Image image = imageService.getImageById(imageId);
+
+        //Calls the comment Service to get all the comments related to an image & return a list.
+        List <Comment> comments = commentService.getCommentsByImageId(imageId);
+
         model.addAttribute("image", image);
-        //model.addAttribute("id", imageId); //attribute added
+        model.addAttribute("comments", comments); //attribute added
         model.addAttribute("tags", image.getTags());
         return "images/image";
     }
@@ -96,13 +105,33 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    //public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session) { //parameter session included
         Image image = imageService.getImage(imageId);
+
+        //Error message introduced
+        String error = "Only the owner of the image can edit the image";
+
+        //Calls the commentService to get all the comments related to an image & return a list.
+
+
+        List <Comment> comments = commentService.getCommentsByImageId(imageId);
 
         String tags = convertTagsToString(image.getTags());
         model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
+        //user defined
+        User user = (User) session.getAttribute("loggeduser");
+
+        //if condition to check for if the loggedUser is the owner
+        if (image.getUser().getId() == user.getId()) {
+            model.addAttribute("tags", tags);
+            return "images/edit";
+        } else {
+            model.addAttribute("editError", error);
+            model.addAttribute("comments", comments);
+            model.addAttribute("tags", image.getTags());
+            return "images/image";
+        }
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -136,7 +165,7 @@ public class ImageController {
         updatedImage.setDate(new Date());
 
         imageService.updateImage(updatedImage);
-        return "redirect:/images/" + updatedImage.getTitle();
+        return "redirect:/images/" + updatedImage.getId() +"/" + updatedImage.getTitle();
     }
 
 
@@ -144,9 +173,25 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    //public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, Model model, HttpSession session) { //parameters model, session included
+        //error message included
+        String error = "Only the owner of the image can delete the image";
+
+        Image image = imageService.getImage(imageId);
+        model.addAttribute("image", image);
+        User user = (User) session.getAttribute("loggeduser");
+        List <Comment> comments = commentService.getCommentsByImageId(imageId);
+
+        if(image.getUser().getId()==user.getId()) {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        } else {
+            model.addAttribute("deleteError", error);
+            model.addAttribute("tags", image.getTags());
+            model.addAttribute("comments", comments);
+            return "images/image";
+        }
     }
